Add set status mutation
