+++ src/schema/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./User";
+export * from "./UserStatus";
